refactor(inventory): 전사재고 조회 영역을 InventoryQueryService 로 분리
InventoryService 1408줄이 전사재고 조회·순환재고·인벤토리 hook·재고
불균형 통계 4개 영역 책임 혼재로 한 파일 1000줄 가이드를 넘어섰음.
본인 영역인 전사재고 조회(findCompanyWide/findCompanyWideSkuDetails)
2메서드와 전용 inner class(Aggregate/SkuAggregate), 헬퍼
(buildLocationOptions/toUiStatus) 를 별 서비스로 분리해 후속 QueryDSL
마이그레이션 변경 범위를 한정한다.

n() 헬퍼는 순환재고 영역에서도 호출되어 다른 팀원 코드 무손상을 위해
새 서비스에 private 으로 중복 작성하고 기존 정의는 그대로 둔다.
Controller 는 InventoryQueryService 를 추가 주입하고 전사재고 2개
엔드포인트 호출 대상만 변경, 순환재고 4개 엔드포인트는 기존
inventoryService 그대로다.
